refactor(service): move definition to repository
