--- domain/definition/consumer.go
@@ -1,7 +1,0 @@
-package definition
-
-import "context"
-
-type ConsumerUsecase interface {
-	DoPayment(ctx context.Context) error
-}
