--- domain/definition/cron.go
@@ -1,7 +1,0 @@
-package definition
-
-import "context"
-
-type CronUsecase interface {
-	DoCreateDummyData(ctx context.Context) error
-}
