+++ domain/repository/mocks/GrpcUsecase.go
@@ -6,7 +6,6 @@ import (
 	context "context"
 
 	customError "github.com/alhamsya/boilerplate-go/lib/helpers/custom_error"
-
 	mock "github.com/stretchr/testify/mock"
 
 	service "github.com/alhamsya/boilerplate-go/protos"
