+++ transport/inter/consumer/routers/types.go
@@ -1,7 +1,7 @@
 package consumerRouters
 
 import (
-	"github.com/alhamsya/boilerplate-go/domain/definition"
+	"github.com/alhamsya/boilerplate-go/domain/repository"
 	"github.com/alhamsya/boilerplate-go/lib/helpers/config"
 )
 
@@ -11,5 +11,5 @@ type ConsumerServer struct {
 }
 
 type ConsumerInteractor struct {
-	ConsumerInterface definition.ConsumerUsecase
+	ConsumerInterface repository.ConsumerUsecase
 }
