+++ transport/inter/grpc/routers/types.go
@@ -1,7 +1,7 @@
 package grpcRouters
 
 import (
-	"github.com/alhamsya/boilerplate-go/domain/definition"
+	"github.com/alhamsya/boilerplate-go/domain/repository"
 	"github.com/alhamsya/boilerplate-go/lib/helpers/config"
 	"google.golang.org/grpc"
 )
@@ -13,5 +13,5 @@ type GrpcServer struct {
 }
 
 type GrpcInteractor struct {
-	GrpcInterface definition.GrpcUsecase
+	GrpcInterface repository.GrpcUsecase
 }
