+++ transport/inter/rest/routers/types.go
@@ -1,7 +1,7 @@
 package restRouters
 
 import (
-	"github.com/alhamsya/boilerplate-go/domain/definition"
+	"github.com/alhamsya/boilerplate-go/domain/repository"
 	"github.com/alhamsya/boilerplate-go/lib/helpers/config"
 	"github.com/alhamsya/boilerplate-go/middleware/rest"
 	"github.com/gofiber/fiber/v2"
@@ -14,6 +14,6 @@ type RestServer struct {
 }
 
 type RestInteractor struct {
-	Usecase    definition.RestUsecase
+	Usecase    repository.RestUsecase
 	Middleware *restMiddleware.Middleware
 }
